# docs/api.yml
openapi: 3.0.0
info:
  title: HospitalityPlatform API
  version: 1.0.0
  description: >-
    The API documentation for HospitalityPlatform, a cross-platform customer
    service and hotel management platform. <br><br>Please note that all API
    documentation will be done with a local instance of Swagger, will use
    OpenAPI 3.0.0, and [can be found on our
    GitHub](https://github.com/COP4331-Group-19/HospitalityPlatform).
  x-logo:
    url: ''
servers:
  - url: https://hospitalityplatform.herokuapp.com/api
    description: Production
  - url: https://virtserver.swaggerhub.com/jeffreydivi/HospitalityPlatform/1.0.0
    description: SwaggerHub API Auto Mocking
paths:
  /account/:
    get:
      tags:
        - Account
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: 'Get current account information and settings.'
          x-last-modified: 1625106621162
      operationId: get_account_settings
      summary: Get current account information and settings.
    patch:
      requestBody:
        description: Edit the current account's settings.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditUser'
        required: false
      tags:
        - Account
      operationId: patch_account_settings
      summary: Edit the current account's settings.
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          x-last-modified: 1625107207680
          description: Edit the current account's settings.
          headers: {}
          links: {}
  /account/letmein/{phone}:
    get:
      parameters:
        - name: phone
          in: path
          description: Phone number
          required: true
          schema:
            type: string
      security: []
      tags:
        - Account
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error200'
          description: 'Reset a password, given a phone number.'
      operationId: get_account_letmein_by_phone
      summary: Reset a password, given a phone number.
  /account/login/:
    post:
      security: []
      summary: Create a session.
      requestBody:
        description: Account credentials
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
        required: true
      tags:
        - Account
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewJWT'
          description: Return the currently logged-in user's account information.
      operationId: post_account_login
      description: Given a username and password, returns a JWT for the session.
...
  /account/create/:
    post:
      tags:
        - Admin
      operationId: post_account_create
      summary: Provision a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserStub'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Provision a new user.
  /hotel/:
    get:
      summary: Get the hotel's metadata.
      tags:
        - Admin
      operationId: get_hotel
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
          description: Returns the hotel's metadata.
    patch:
      tags:
        - Admin
      operationId: patch_hotel
      summary: Edit the instance's metadata.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Hotel'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
          description: Returns the hotel's metadata.
...
  /room/{room_id}:
    get:
      tags:
        - Admin
      operationId: get_room_by_room_id
      summary: Get room information by room number.
      parameters:
        - name: room_id
          in: path
          description: Room number
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
          description: Get room information by room number.
    post:
      tags:
        - Admin
      operationId: post_room_by_room_id
      summary: Create a room with the given room number, if it does not yet exist.
      parameters:
        - name: room_id
          in: path
          description: Room number
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
          description: Returns the new room's information.
    patch:
      tags:
        - Admin
      operationId: patch_room_by_room_id
      summary: Edit a preexisting room.
      parameters:
        - name: room_id
          in: path
          description: Room number
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
          description: Returns the updated room's information.
    delete:
      tags:
        - Admin
      operationId: delete_room_by_room_id
      summary: Deletes a room.
      parameters:
        - name: room_id
          in: path
          description: Room number
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error200'
          description: Returns a success message of the room's deletion.
  /floor/{floor_number}:
    get:
      tags:
        - Admin
      operationId: get_floor_by_floor_number
      summary: Get all rooms on a given floor.
      parameters:
        - name: floor_number
          in: path
          description: Floor number
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
          description: Returns all rooms on a given floor.
  /inventory/:
    get:
      tags:
        - Guest
      operationId: get_inventory
      summary: List all inventory items.
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Inventory'
          description: List all inventory items.
    post:
      tags:
        - Admin
      operationId: post_inventory
      summary: Add an item to the inventory.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditInventory'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
          description: Returns the new inventory item's information.
  /inventory/{inventory_id}/{quantity}:
    get:
      tags:
        - Guest
      operationId: get_inventory_by_inventory_id_with_quantity
      parameters:
        - name: inventory_id
          in: path
          description: Inventory item's ID
          required: true
          schema:
            type: integer
        - name: quantity
          in: path
          description: Amount of items to order.
          required: true
          schema:
            type: integer
      summary: Orders an inventory item to a user's room.
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: Orders an inventory item to a user's room.
  /inventory/{inventory_id}:
    get:
      tags:
        - Staff
        - Guest
      operationId: get_inventory_by_inventory_id
      summary: Get information on a specific inventory entry.
      parameters:
        - name: inventory_id
          in: path
          description: Inventory item's ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
          description: Returns the requested inventory entry.
    patch:
      tags:
        - Admin
      operationId: patch_inventory_by_inventory_id
      summary: Edit a preexisting inventory item.
      parameters:
        - name: inventory_id
          in: path
          description: Inventory item's ID
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Inventory'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
          description: Returns the updated inventory entry's information.
    delete:
      tags:
        - Admin
      operationId: delete_inventory_by_inventory_id
      summary: Deletes an inventory item.
      parameters:
        - name: inventory_id
          in: path
          description: Inventory item's ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error200'
          description: Returns a success message of the inventory item's deletion.
...
components:
  schemas:
...
    Error200:
      required:
        - err_code
        - description
      type: object
      properties:
        err_code:
          description: A numeric error code.
          type: integer
          example: 200
        description:
          description: A human-readable explanation of the error.
          type: string
          example: Success!
...
    User:
      description: A complete single user entry.
      allOf:
        - required:
            - user_id
            - first_name
            - username
            - password
            - role
          type: object
          properties:
            user_id:
              description: The user ID.
              type: integer
              example: 7
            role:
              description: The role of the account. Either admin, staff, or guest.
              type: string
              example: admin
            checkin:
              description: The currently active check-in date and time, stored in Unix time.
              type: integer
              example: 1625184007928
            checkout:
              description: The currently active check-out date and time, stored in Unix time.
              type: integer
              example: 1625184007929
            room:
              description: The currently active room number.
              type: string
              example: 345P
        - $ref: '#/components/schemas/EditUser'
...
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
...
security:
  - Bearer: []
...